Use protobuf argument for Client request methods
Change the argument for Client API request methods to be the
corresponding protobuf request message. This adds some boilerplate when
calling this methods, but allows complete flexibility in setting the
request parameters.

Also rename these methods to use snake_case instead of matching the name
of the API in the URL.

Closes #140.