Fix 403 error by adding API key parameter
Add API key parameter to requests to avoid new error that started today:
403 Forbidden "Daily Limit for Unauthenticated Use Exceeded. Continued
use requires signup."

Older versions of hangups sent the key parameter, but it was later
removed because it wasn't necessary and removing it allowed the
connection process to be simplified.

The API key AIzaSyAfFJCeph-euFSwtmqFZi0kaKk-cZ5wufM has been the same
for a long time, so hard code it for now.

Fixes #265.