Obtain the initial conversation list more robustly
As it turns out, Hangouts can create conversations without a single
event (including invites to the conversation itself). Also as it turns
out, there are three separate locations where a timestamp is obtainable
in the ConversationState proto:

- `conversation.self_conversation_state.sort_timestamp`
- `event_continuation_token.event_timestamp`
- `event[0].timestamp`

Switch from using `event[0].timestamp` to
`conversation.self_conversation_state.sort_timestamp` to retrieve
pages of conversations:

- SyncRecentConversations appears to monotonically order returned
  conversations by their `sort_timestamp`, NOT their event timestamps
- Hangups uses the sort timestamp for the `last_modified` conversation
  attribute, which gets called from the UI

Upon testing, this appeared to return a maximally large set of hangouts,
so it's a more correct method than the existing use of the event
timestamps anyway.

Also, update `examples/build_conversation_list.py`, as the updated code
was useful in tracking down a potential solution to #297.