Use latest version of aiohttp.
The requires line for aiohttp was previously using <2,
which caused a old 2.0 release candidate version of aiohttp
to be installed. It now uses <3.

Additionally, >=2 is now used as the minimum aiohttp version.
I found anything <2 is buggy when it comes to timeout handling;
Calls like request() would just hang indefinitely, despite a
connection timeout being provided. I believe that this was the
source of issue #301.

Also fixed a few errors like this:
"ClientError is an ancestor class of ServerDisconnectedError."
These probably cropped up thanks to the newer version of aiohttp.